Added GET endpoint + refactoring.
