+++ .env.example
@@ -1,0 +1,1 @@
+PROJECT_VERSION = 0.1
