+++ .gitignore
@@ -1,3 +1,4 @@
+# Temporary folder
 tmp/
 
 # Byte-compiled / optimized / DLL files
@@ -160,3 +161,6 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
+
+# VS Code
+.vscode/
