+++ docker-compose.yml
@@ -7,7 +7,7 @@ services:
     build: ./mocker-service
     command: uvicorn main:app --reload --host 0.0.0.0 --port 8000
     volumes:
-      - ./mocker-service/app:/app
+      - ./mocker-service/mocker:/app
     expose:
       - 8000
