--- mocker-service/app/test_main.py
@@ -1,13 +1,0 @@
-from fastapi.testclient import TestClient
-
-from .main import app
-
-
-client = TestClient(app)
-
-
-def test_read_main():
-    response = client.get("/")
-    assert response.status_code == 200
-    assert response.headers["content-type"] == "application/json"
-    assert response.json() == "Placeholder!"
