+++ mocker-service/mocker/__init__.py
@@ -1,0 +1,1 @@
+"""Mocker application package."""
