+++ pyproject.toml
@@ -1,4 +1,4 @@
 [tool.isort]
 profile = "black"
-src_paths = ["mocker-service/*"]
+src_paths = ["mocker-service/mocker"]
 lines_after_imports = 2
